requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
blis==0.7.11
build==1.2.1
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
debugpy==1.8.5
decorator==5.1.1
django-debug-toolbar==4.4.6
django-seed==0.3.1
dnspython==2.6.1
executing==2.0.1
Faker==27.0.0
flexcache==0.3
flexparser==0.3.1
fonttools==4.53.1
greenlet==3.0.3
idna==3.7
ipykernel==6.29.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mathparse==0.1.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.10
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
Pint==0.24.3
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==3.3.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
PyYAML==3.10
pyzmq==26.1.0
regex==2024.7.24
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
seaborn==0.13.2
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.2.0
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
thinc==8.2.5
threadpoolctl==3.5.0
toposort==1.10
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wheel==0.44.0
wrapt==1.16.0
